enchant: Autospec creation for update from version 2.6.5 to version 2.6.6
Reuben Thomas (7):
      Makefile.am: in release target, push commits and tags in one command
      Makefile.am: in release target, checkout doxygen directory to avoid spurious diffs
      NEWS: add a note for the next version about DICPATH (see #354)
      Add more debug logging, mainly for attempts to read config files
      Fix paths used for reading configuration files
      configure.ac: bump version to 2.6.6
      Add NEWS for 2.6.6

2.6.6 (February 4, 2024)
------------------------

This version fixes the implementation and documentation of system paths for
configuration files (enchant.ordering). Since version 2.6.4, the wrong path
was searched for pkgdatadir: not, for example /usr/share/enchant-2, as
documented, but the old path /usr/share/enchant. The enchant(5) man page
referred to “DATADIR” rather than the actual configured path. The sysconfdir
location, typically /etc, was not documented.